test: fix card tests and coverage
