+++ packages/card/__tests__/ios/card.ios.test.js
@@ -1,5 +1,5 @@
-import cardTests from "../card.native.test.js";
+import shared from "../shared";
 
 describe("Card tests on ios", () => {
-  cardTests();
+  shared();
 });
